refactor: 컴포넌로트 분리
